refactor: co-locate builders with solvers to break import cycles (#825)
The latent cla<->builder and lasso<->builder import cycles were previously
broken only at runtime via TYPE_CHECKING guards plus function-local imports.
Co-locate each fluent builder with the object it constructs so the internal
import graph is genuinely acyclic:

- move ProblemBuilder into cla.py (next to CLA)
- move LassoBuilder into lasso.py (next to Lasso)
- reduce builder.py to a thin re-export facade preserving the historical
  cvxcla.builder.{ProblemBuilder,LassoBuilder} import path
- drop the now-unnecessary TYPE_CHECKING/function-local builder imports from
  the solver modules

The solver modules no longer import builder at all; builder imports the
solvers one way. Rename tests/test_deferred_imports.py -> test_import_graph.py
and rewrite it to pin the new acyclic direction. Behaviour, public API, and the
100% coverage gate are unchanged (200 passed).
