+++ src/cvxcla/lasso.py
@@ -46,7 +46,7 @@ from __future__ import annotations
 from dataclasses import dataclass, field
 from functools import cached_property
 from itertools import pairwise
-from typing import TYPE_CHECKING, cast
+from typing import cast
 
 import numpy as np
 from numpy.typing import NDArray
@@ -55,9 +55,6 @@ from ._lasso import LassoSegment, LassoState, scan_events, solve_segment
 from ._lasso_validate import validate_constraints, validate_design_inputs, validate_operator_inputs
 from .operators import DenseCovariance, GramCovariance, QuadraticForm
 from .pathtracer import InequalityConstrained, trace
-
-if TYPE_CHECKING:
-    from .builder import LassoBuilder
 
 
 @dataclass(frozen=True)
@@ -155,12 +152,6 @@ class Lasso(InequalityConstrained):
         Returns:
             A :class:`cvxcla.builder.LassoBuilder`.
         """
-        # Deferred (function-local) import on purpose: ``builder`` imports ``Lasso`` at
-        # module level, so importing it at the top here would form a real import cycle.
-        # This factory only needs ``LassoBuilder`` at call time, so the import is safe to
-        # defer. See also the ``TYPE_CHECKING`` import above and ``CLA.problem``.
-        from .builder import LassoBuilder
-
         return LassoBuilder(x, y)
 
     @classmethod
@@ -353,3 +344,90 @@ class Lasso(InequalityConstrained):
                 return (1.0 - weight) * lo.beta + weight * hi.beta
         msg = "lam lies within the path range but no bracketing segment was found"  # pragma: no cover
         raise AssertionError(msg)  # pragma: no cover
+
+
+class LassoBuilder:
+    """Chainable builder for a LASSO regularisation-path problem.
+
+    The LASSO counterpart of :class:`cvxcla.cla.ProblemBuilder`. Construct one via
+    :meth:`Lasso.problem`, optionally add inequality constraints with
+    :meth:`inequality`, and finish with :meth:`trace`, which builds the
+    :class:`Lasso` and traces the entire regularisation path. Like the CLA builder
+    it adds no modelling power: it accepts the same ``G beta <= h`` rows the
+    ``Lasso`` already supports and nothing else.
+
+    Examples:
+        >>> import numpy as np
+        >>> from cvxcla import Lasso
+        >>> rng = np.random.default_rng(0)
+        >>> x = rng.standard_normal((30, 5))
+        >>> y = rng.standard_normal(30)
+        >>> lasso = Lasso.problem(x, y).trace()
+        >>> len(lasso.path) > 0
+        True
+    """
+
+    def __init__(self, x: NDArray[np.float64], y: NDArray[np.float64]) -> None:
+        """Start a builder for design matrix ``x`` and response ``y``.
+
+        Args:
+            x: Design matrix of shape ``(m, n)``.
+            y: Response vector of shape ``(m,)``.
+        """
+        self.x = np.asarray(x, dtype=np.float64)
+        self.y = np.asarray(y, dtype=np.float64)
+        self._g_blocks: list[NDArray[np.float64]] = []
+        self._h_blocks: list[NDArray[np.float64]] = []
+        self._nonneg = False
+
+    def non_negative(self) -> LassoBuilder:
+        """Restrict the coefficients to ``beta >= 0`` (the non-negative LASSO).
+
+        Under ``beta >= 0`` the l1 penalty collapses to the linear term
+        ``lam * sum(beta)``, so the path is the standard one restricted to positive
+        signs -- structurally the CLA's box-bounded parametric QP.
+
+        Returns:
+            ``self``, for chaining.
+        """
+        self._nonneg = True
+        return self
+
+    def inequality(self, g: NDArray[np.float64], h: float | NDArray[np.float64]) -> LassoBuilder:
+        """Add one or more inequality rows ``G beta <= h`` (repeated calls accumulate).
+
+        Args:
+            g: A length-``n`` row vector or a ``(p, n)`` matrix.
+            h: The matching right-hand side: a scalar for a single row, or a
+                length-``p`` vector. Each entry must be strictly positive (so
+                ``beta = 0`` stays feasible), checked when the path is traced.
+
+        Returns:
+            ``self``, for chaining.
+
+        Raises:
+            ValueError: If ``g``'s column count is not ``n`` or ``h``'s length does
+                not match the rows of ``g``.
+        """
+        g_block = np.atleast_2d(np.asarray(g, dtype=np.float64))
+        h_block = np.atleast_1d(np.asarray(h, dtype=np.float64))
+        if self.x.ndim == 2 and g_block.shape[1] != self.x.shape[1]:
+            msg = f"inequality: coefficient matrix must have {self.x.shape[1]} columns, got shape {g_block.shape}"
+            raise ValueError(msg)
+        if h_block.shape[0] != g_block.shape[0]:
+            msg = f"inequality: h must have {g_block.shape[0]} entries to match the rows, got {h_block.shape[0]}"
+            raise ValueError(msg)
+        self._g_blocks.append(g_block)
+        self._h_blocks.append(h_block)
+        return self
+
+    def trace(self) -> Lasso:
+        """Assemble the pieces, build the ``Lasso``, and trace the full path.
+
+        Returns:
+            The traced :class:`Lasso`, whose ``path`` holds the breakpoints of the
+            (constrained) regularisation path.
+        """
+        g = np.vstack(self._g_blocks) if self._g_blocks else None
+        h = np.concatenate(self._h_blocks) if self._h_blocks else None
+        return Lasso(x=self.x, y=self.y, g=g, h=h, nonneg=self._nonneg)
